requirements.txt
smbus
spidev
gpiozero
pillow
lgpio